Consolidate daily sync into acms-cams-transition
Moves all CMMAP_ALL and CMMAP_SYNC_CONTROL logic into the single
acms-cams-transition module so all downstream ACMS transition
handlers share one file, one connection pool, and one test suite.
acms-daily-sync.ts is now a thin re-export shim.

Jira ticket: CAMS-616
